1st Commit
If I can get the game to upload it'll be here, but I am not very capable so it will be very disappointing and a showcase of my sheer inadequacy. However, I love to write, so please enjoy the further point story I wrote! Its not too deep, but the game isn't either so I figure its quite fair.